requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.0
SQLAlchemy>=2.0.30
pytest==8.1.1
requests>=2.28.0
openai>=1.0.0
openai-whisper>=20231117
